server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
faiss-cpu==1.7.4
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.5
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.0.2
openai==1.95.0
orjson==3.10.18
packaging==24.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.13.2
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dotenv==1.1.1
python-magic==0.4.27
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
